Implementation Of The Collision System In Multiple Layers 1. The Collider Class Has A Copy Of The Layer Maps From The Current Level. 2. Fixed Some Parsing Problems In Multiple Classes. 3. Updated The .tmx From The SLevelOneJapan Level. PD: There Are Some Bugs That Soon Will Be Fixed.

+++ ModelingProject1/SourceCode/Camera.cpp
@@ -121,7 +121,7 @@ void Camera::restartCamera(GLfloat level)
 
 bool Camera::isPlayerOnMidpoint(GLfloat posX)
 {
-  int distanceToMid = posX - midPoint;
+  int distanceToMid = (int)posX - (int)midPoint;
   if(playerOnMidpoint)
   {
     if(abs(distanceToMid) < 250)
